fix: 싫어요 중복 가능, 싫어요를 누르지 않았던 쓰레기통 싫어요 취소 가능한 문제 해결

+++ src/main/java/net/binder/api/memberdislikebin/service/MemberDislikeBinService.java
@@ -3,6 +3,7 @@ package net.binder.api.memberdislikebin.service;
 import lombok.RequiredArgsConstructor;
 import net.binder.api.bin.entity.Bin;
 import net.binder.api.bin.service.BinService;
+import net.binder.api.common.exception.BadRequestException;
 import net.binder.api.member.entity.Member;
 import net.binder.api.member.service.MemberService;
 import net.binder.api.memberdislikebin.entity.MemberDislikeBin;
@@ -23,6 +24,11 @@ public class MemberDislikeBinService {
 
     public void saveDislike(String email, Long binId){
         Member member = memberService.findByEmail(email);
+
+        if (memberDislikeBinRepository.existsByMember_IdAndBin_Id(member.getId(), binId)) {
+            throw new BadRequestException("이미 싫어요를 누른 쓰레기통입니다.");
+        }
+
         Bin bin = binService.findById(binId);
         bin.increaseDislike();
         MemberDislikeBin memberDislikeBin = MemberDislikeBin.builder()
@@ -33,6 +39,11 @@ public class MemberDislikeBinService {
     }
 
     public void deleteDisLike(String email, Long binId){
+        Member member = memberService.findByEmail(email);
+        if (!memberDislikeBinRepository.existsByMember_IdAndBin_Id(member.getId(), binId)) {
+            throw new BadRequestException("싫어요를 누르지 않았던 쓰레기통입니다.");
+        }
+
         Bin bin = binService.findById(binId);
         bin.decreaseDisLike();
         memberDislikeBinRepository.deleteMemberLikeBinByMember_EmailAndBin_Id(email, binId);
